More work on 16.
For each position, consider whether we've been in this position before, but with more time left and more valves open. If so, don't bother exploring from here.

I think it's a coincidence; this fails for the example, but works for the input. I think the example has a weird case where I'm discarding a 3-valve, fewer-points state to a 4-valve, more-points state, but the 3-valve state would overtake the 4-valve state by opening another valve. If that makes sense.