Day 19 sim %
Spent the day making a neat little interactive RTS instead of trying to solve the problem, which kind of just looks like it's going to be a BFS of every possible game action per turn.

Ah, I bet you'll be able to build multiple robots on the same turn. Dang.